+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+* Add support for `textDocument/prepareRename` request;
 * Upgrade Microsoft.Build + Roslyn libs;
 * update textDocument/documentSymbol handler:
   - emit hierarchical symbol tree, where client supports it;