+++ CHANGELOG.md
@@ -3,12 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.5.7] - 2022-10-09 / Marijampolė
 * Add support for `textDocument/prepareRename` request;
 * Upgrade Microsoft.Build + Roslyn libs;
-* Update textDocument/documentSymbol handler:
+* Update `textDocument/documentSymbol` handler:
   - emit hierarchical symbol tree, where client supports it;
-  - emit DocumentSymbol[] instead of SymbolInformation[], the former of which has more info.
+  - emit `DocumentSymbol[]` instead of `SymbolInformation[]`, the former of which has more info.
+
+### More about Marijampolė, Lithuania
+- [Google Images on Marijampolė](https://www.google.com/search?tbm=isch&q=marijampole+lithuania)
+- [Marijampolė on WP](https://en.wikipedia.org/wiki/Marijampol%C4%97)
 
 ## [0.5.6] - 2022-09-05 / Birštonas
 * Fix an issue where we would fail to generate code action list if one of code action providers would raise an exception; -- log the message to the log instead and continue with the next provider to be more resilent;